- Mejorada la implementación inicial de la intranet documental. - Desarrollado e implementado el index empleado. - Se ha añadido un permiso denominado control total y se le ha asignado al rol administrador - Se ha implementado como condición para poder acceder al menu de ajustes que el usuario autentificado sea administrador y por ende tenga el permiso control total. - Se han mejorado las relaciones de las entidades user y centro - Se ha reajustado el orden de los elementos en el menu del dashboard - Ahora en la página del dashboard no solo se indica el usuario autentificado y el rol del mismo, tambien se indica el centro productivo al que pertenece y por ende desde el cual esta trabajando. - Se han movido las llamadas a las alertas de sweet alert del componente de livewire a las vistas pertinentes - Se ha implementado la ficha del empleado y la edición de la misma incluyendo  rutas, validaciones y con las alertas de sweet alert implementadas. - Se ha implementado la API de openstreetmap en la ficha del empleado, para asi de esta manera geocalizar en un mapa a cada empleado según su dirección, población , país etc

+++ app/Http/Controllers/CentroController.php
@@ -110,10 +110,7 @@ class CentroController extends Controller
             ]);
     
 
-                //creando directorios raiz de carpetas para cada centro
-                Storage::disk('local')->makeDirectory('intranet/'.$request->input('nombre'));
-                Storage::disk('local')->makeDirectory('intranet/'.$request->input('nombre').'/empleados');
-                Storage::disk('local')->makeDirectory('intranet/'.$request->input('nombre').'/proyectos');
+
 
                 $centro = new Centro();
                 $centro->nombre = $request->input('nombre');
@@ -291,14 +288,29 @@ class CentroController extends Controller
     {
 
         try{
-            if(PermisosController::authAdmin()){        
-                $centro = Centro::find($id);
-                //borrando directorio de carpetas
-                Storage::disk('local')->deleteDirectory('intranet/'.$centro->nombre);
-                $centro->delete();
-                session()->flash('eliminado', 'ok');
 
-                return back();
+            $centros = Centro::all();
+
+
+            if(PermisosController::authAdmin()){        
+
+                if(count($centros) == 1){
+
+                    return view('Mensaje.advertencia', ['titulo' => 'Operación no disponible', 'mensaje' => 'Solo existe un centro de trabajo y no puede ser eliminado.']);
+
+                }else{
+
+                    $centro = Centro::find($id);
+                    //borrando directorio de carpetas
+                    Storage::disk('local')->deleteDirectory('intranet/'.$centro->nombre);
+                    $centro->delete();
+                    session()->flash('eliminado', 'ok');
+    
+                    return back();
+                    
+                }
+
+    
             }else{
 
                 return view('Mensaje.advertencia', ['titulo' => 'Operación no disponible', 'mensaje' => 'Este usuario no puede eliminar un Centro Productivo. Pongase en contacto con su administrador.']);
